themix: exit themix run after finish
add type STATUS to indicate the status of themix

use statusCh to inform themixQue this themix is finish and can set its related queue status to TOMBSTONE, which informs that messages in this sequence will not be handled and close msgc in for this themix.

msgc is changed from unbufferred to buffer channel, which solves dead clock when both side use select-for loop.